+++ lib/protobuf/rpc/servers/zmq/worker.rb
@@ -1,5 +1,6 @@
 require 'protobuf/rpc/server'
 require 'protobuf/rpc/servers/zmq/util'
+require 'thread'
 
 module Protobuf
   module Rpc
@@ -51,8 +52,10 @@ module Protobuf
             break if rc == -1
 
             if rc > 0
+              ::Thread.current[:busy] = true
               initialize_request!
               process_request
+              ::Thread.current[:busy] = false
             end
           end
         ensure
@@ -74,12 +77,20 @@ module Protobuf
         private
 
         def init_zmq_context
-          @zmq_context = ZMQ::Context.new
+          if inproc?
+            @zmq_context = @server.zmq_context
+          else
+            @zmq_context = ZMQ::Context.new
+          end
         end
 
         def init_backend_socket
           @backend_socket = @zmq_context.socket(ZMQ::REQ)
           zmq_error_check(@backend_socket.connect(@server.backend_uri))
+        end
+
+        def inproc?
+          !!@server.try(:inproc?)
         end
 
         def read_from_backend