+++ CHANGES.md
@@ -5,12 +5,24 @@ Unreleased (3.x)
 - Clean up the Enum class API with several methods deprecated. See #134 for more comprehensive
 documentation about which methods are going and away and which are being renamed. [#134, @localshred]
 
-
 --------
-
 
 Stable
 --------
+
+2.8.12
+---------
+
+- Fix thread busy access in zmq server/worker. [#151, @abrandoned]
+
+2.8.11
+---------
+
+- Default ZMQ server to use inproc protocol instead of tcp (zero-copy between server-broker-worker). [#145, @brianstien]
+- Add `broadcast_busy` functionality that removes server from cluster if the workers are full. [#149, @abrandoned]
+- Add cli option for `--no-zmq_inproc`. [#149, @abrandoned]
+- Add cli option for `--broadcast_busy`. [#149, @abrandoned]
+>>>>>>> 2-8-stable
 
 2.8.10
 ---------